+++ src/experimental/base/array/__tests__/ArrayComponent.test.ts
