+++ src/experimental/base/array/__tests__/fixtures/comp1.ts
