+++ src/experimental/base/array/__tests__/fixtures/comp2.ts
