+++ src/experimental/base/index.ts
@@ -3,4 +3,4 @@ export { Component } from './component/Component';
 export { NestedComponent } from './nested/NestedComponent';
 export { DataComponent } from './data/DataComponent';
 export { ArrayComponent } from './array/ArrayComponent';
-export * from 'model';
+export * from 'src/experimental/model';
