+++ src/experimental/base/nested/__tests__/fixtures/comp1.ts
