+++ src/experimental/base/nested/__tests__/fixtures/comp2.ts
