+++ src/experimental/components/__tests__/fixtures/comp1.ts
