+++ src/experimental/components/__tests__/fixtures/comp2.ts
