+++ src/index.ts
@@ -1,9 +1,5 @@
-export * from './core';
-export * from './base';
-export * from './model';
 export * from './modules';
 export * from './utils';
-export * from './components';
 export * from './process/validation';
 export * from './process';
-export * from './template';
+export * from './sdk';
